Added $_SESSION mechanism and fixed bug in validation

+++ delete.php
@@ -1,12 +1,18 @@
 <?php
-include_once("Class/Crud.php");
+include_once("Class/Crud.php"); 
 
 $crud = new Crud();
-
-$id= $_GET['id'];
-$stmt = $crud->delete($id);
-
-if ($stmt) {
+session_start();
+if(!isset($_SESSION["sessionVar"])) {
     header("Location:table.php");
-}   
+} else {
+
+    $id= $_GET['id'];
+    $stmt = $crud->delete($id);
+
+    if ($stmt) {
+        header("Location:table.php");
+        }
+    } 
+    
 ?>
